# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel REST API V4 - Master-Detail Architecture",
        "description": "REST API untuk Angular Frontend dengan Master-Detail Architecture (Many-to-Many)<br><br>\n *                  <strong>Fitur V4:</strong><br>\n *                  ✅ Business Units & Menus sebagai Master Data<br>\n *                  ✅ Many-to-Many relationships (User ↔ BU, User ↔ Menu)<br>\n *                  ✅ Dropdown API untuk manage user access<br>\n *                  ✅ Dynamic menu & BU assignment<br><br>\n *                  <strong>Test Accounts:</strong><br>\n *                  - admin / Admin123 (Full access)<br>\n *                  - user1 / User123 (Batam, Jakarta + 3 menus)<br>\n *                  - user2 / User123 (Surabaya + 2 menus)",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/users/{id}/access": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user access data for dropdown (V4 Critical API)",
                "description": "Returns user info + accessible business units + menus in one call. Perfect untuk populate dropdown di Angular.",
                "operationId": "5fc3fc2ce087c8f565b42a0013480d2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "username": {
                                                            "type": "string",
                                                            "example": "user1"
                                                        },
                                                        "level": {
                                                            "type": "string",
                                                            "example": "user"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "business_units": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "business_unit": {
                                                                "type": "string",
                                                                "example": "Batam"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "menus": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nama_menu": {
                                                                "type": "string",
                                                                "example": "Dashboard"
                                                            },
                                                            "url_link": {
                                                                "type": "string",
                                                                "example": "/dashboard"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/business-units": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Assign business units to user",
                "description": "Assign array of business unit IDs to user. This will REPLACE existing assignments.",
                "operationId": "bfdff9f00c190626684d7abae2c2c8f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "business_unit_ids"
                                ],
                                "properties": {
                                    "business_unit_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business units assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business units updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/menus": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Assign menus to user",
                "description": "Assign array of menu IDs to user. This will REPLACE existing assignments.",
                "operationId": "2b6ce7fc8832919f85d890010eed59ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "menu_ids"
                                ],
                                "properties": {
                                    "menu_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            7
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menus assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Menus updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}